version fusionnée (#20)
* 2ème version evaluateur et test evaluateur (#2)

* 2ème version évaluateur et test évaluateur

---------


* Add files via upload

* Add files via upload

* Add files via upload

* version finale fusion gogeta

---------

+++ evaluateur.h
@@ -3,8 +3,9 @@
 
 #include <math.h>
 #include <stdlib.h>
+#include <stdio.h>
 #include "jeton.h"
 
-float Eval( Arbre A,int x,int *erreur_pg);
+float Eval(Arbre A, float x, int* erreur_pg);
 
 #endif
